pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crawl-telegraph"
version = "0.1.0"
description = "Small toolset to scrape and download images from telegra.ph pages."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Kerry", email = "kerry@example.com" }
]
requires-python = ">=3.8"
keywords = ["telegra.ph", "crawler", "images", "download"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
    "requests>=2.28,<3",
    "webui2>=2.5.6,<3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<8",
    "black>=24.0,<25"
]

[project.scripts]
crawl-telegraph = "main:main"
crawl-telegraph-ui = "ui:launch_ui"

[project.urls]
Homepage = ""
Repository = ""
Source = "https://example.com/your-repo"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]
